+++ example/res/layout/activity_add_address_example.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
     >
 
     <com.stripe.android.view.AddAddressWidget
@@ -11,4 +10,4 @@
         android:layout_height="wrap_content"
         android:layout_margin="12dp"
         />
-</LinearLayout>
+</FrameLayout>
