+++ stripe/res/layout/activity_add_source.xml
@@ -1,36 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     >
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/add_source_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="@dimen/toolbar_elevation"
-        app:title="@string/title_add_a_card"
-        android:theme="@style/ToolBarStyle"
-        />
 
     <com.stripe.android.view.CardMultilineWidget
         android:id="@+id/add_source_card_entry_widget"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@id/add_source_toolbar"
         android:layout_margin="@dimen/add_card_total_margin"
-        />
-
-    <ProgressBar
-        android:id="@+id/add_source_progress_bar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-        android:indeterminate="true"
-        android:layout_below="@id/add_source_toolbar"
         />
 
     <FrameLayout
