+++ stripe/res/layout/add_address_widget.xml
@@ -8,7 +8,6 @@
         android:id="@+id/spinner_country_aaw"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
         />
 
     <android.support.design.widget.TextInputLayout
