+++ stripe/src/main/java/com/stripe/android/view/AddSourceActivity.java
@@ -2,20 +2,14 @@ package com.stripe.android.view;
 
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.drawable.Drawable;
+import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.VisibleForTesting;
 import android.support.transition.Fade;
 import android.support.transition.TransitionManager;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.FrameLayout;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.stripe.android.CustomerSession;
@@ -28,22 +22,20 @@ import com.stripe.android.model.Source;
 import com.stripe.android.model.SourceParams;
 import com.stripe.android.model.StripePaymentSource;
 
-public class AddSourceActivity extends AppCompatActivity {
+public class AddSourceActivity extends StripeActivity {
 
     public static final String EXTRA_NEW_SOURCE = "new_source";
     static final String ADD_SOURCE_ACTIVITY = "AddSourceActivity";
     static final String EXTRA_SHOW_ZIP = "show_zip";
     static final String EXTRA_UPDATE_CUSTOMER = "update_customer";
-    static final long FADE_DURATION_MS = 100L;
     CardMultilineWidget mCardMultilineWidget;
     CustomerSessionProxy mCustomerSessionProxy;
-    ProgressBar mProgressBar;
     TextView mErrorTextView;
     FrameLayout mErrorLayout;
     StripeProvider mStripeProvider;
-    Toolbar mToolbar;
 
-    private boolean mCommunicating;
+    static final long FADE_DURATION_MS = 100L;
+
     private boolean mUpdatesCustomer;
 
     /**
@@ -68,67 +60,20 @@ public class AddSourceActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_source);
+        mViewStub.setLayoutResource(R.layout.activity_add_source);
+        mViewStub.inflate();
         mCardMultilineWidget = findViewById(R.id.add_source_card_entry_widget);
-        mProgressBar = findViewById(R.id.add_source_progress_bar);
         mErrorTextView = findViewById(R.id.tv_add_source_error);
-        mToolbar = findViewById(R.id.add_source_toolbar);
-        setSupportActionBar(mToolbar);
-        if (getSupportActionBar() != null) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        }
-        setCommunicatingProgress(false);
+        mErrorLayout = findViewById(R.id.add_source_error_container);
         boolean showZip = getIntent().getBooleanExtra(EXTRA_SHOW_ZIP, false);
         mUpdatesCustomer = getIntent().getBooleanExtra(EXTRA_UPDATE_CUSTOMER, false);
         mCardMultilineWidget.setShouldShowPostalCode(showZip);
-
-        mErrorLayout = findViewById(R.id.add_source_error_container);
+        setTitle(R.string.title_add_a_card);
     }
+
 
     @Override
-    public boolean onPrepareOptionsMenu(Menu menu) {
-        MenuItem saveItem = menu.findItem(R.id.action_save);
-        Drawable tintedIcon = ViewUtils.getTintedIcon(
-                this,
-                R.drawable.ic_checkmark,
-                android.R.color.primary_text_dark);
-        saveItem.setIcon(tintedIcon);
-        return super.onPrepareOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.add_source_menu, menu);
-        menu.findItem(R.id.action_save).setEnabled(!mCommunicating);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.getItemId() == R.id.action_save) {
-            saveCardOrDisplayError();
-            return true;
-        } else {
-            boolean handled = super.onOptionsItemSelected(item);
-            if (!handled) {
-                onBackPressed();
-            }
-            return handled;
-        }
-    }
-
-    @VisibleForTesting
-    void setCustomerSessionProxy(CustomerSessionProxy proxy) {
-        mCustomerSessionProxy = proxy;
-    }
-
-    @VisibleForTesting
-    void setStripeProvider(@NonNull StripeProvider stripeProvider) {
-        mStripeProvider = stripeProvider;
-    }
-
-    private void saveCardOrDisplayError() {
+    protected void onActionSave() {
         mErrorTextView.setVisibility(View.GONE);
         Card card = mCardMultilineWidget.getCard();
         if (card == null) {
@@ -194,24 +139,6 @@ public class AddSourceActivity extends AppCompatActivity {
         finish();
     }
 
-    private Stripe getStripe() {
-        if (mStripeProvider == null) {
-            return new Stripe(this);
-        } else {
-            return mStripeProvider.getStripe(this);
-        }
-    }
-
-    private void setCommunicatingProgress(boolean communicating) {
-        mCommunicating = communicating;
-        if (communicating) {
-            mProgressBar.setVisibility(View.VISIBLE);
-        } else {
-            mProgressBar.setVisibility(View.GONE);
-        }
-        supportInvalidateOptionsMenu();
-    }
-
     private void showError(@NonNull String error, boolean shouldAnimate) {
         mErrorTextView.setText(error);
         if (shouldAnimate) {
@@ -220,6 +147,25 @@ public class AddSourceActivity extends AppCompatActivity {
             TransitionManager.beginDelayedTransition(mErrorLayout, fadeIn);
         }
         mErrorTextView.setVisibility(View.VISIBLE);
+    }
+
+    private Stripe getStripe() {
+        if (mStripeProvider == null) {
+            return new Stripe(this);
+        } else {
+            return mStripeProvider.getStripe(this);
+        }
+    }
+
+
+    @VisibleForTesting
+    void setCustomerSessionProxy(CustomerSessionProxy proxy) {
+        mCustomerSessionProxy = proxy;
+    }
+
+    @VisibleForTesting
+    void setStripeProvider(@NonNull StripeProvider stripeProvider) {
+        mStripeProvider = stripeProvider;
     }
 
     interface StripeProvider {
