+++ stripe/src/main/java/com/stripe/android/view/StripeActivity.java
@@ -1,0 +1,91 @@
+package com.stripe.android.view;
+
+import android.graphics.drawable.Drawable;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewStub;
+import android.widget.ProgressBar;
+
+import com.stripe.android.R;
+
+/**
+ * Provides a toolbar, save button, and loading states for the save button.
+ */
+abstract class StripeActivity extends AppCompatActivity {
+
+    boolean mCommunicating;
+    Toolbar mToolbar;
+    ProgressBar mProgressBar;
+    ViewStub mViewStub;
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_stripe);
+        mProgressBar = findViewById(R.id.progress_bar_as);
+        mToolbar = findViewById(R.id.toolbar_as);
+        mViewStub = findViewById(R.id.widget_viewstub_as);
+        if (getSupportActionBar() != null) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        }
+        setSupportActionBar(mToolbar);
+        if (getSupportActionBar() != null) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        }
+        setCommunicatingProgress(false);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.add_source_menu, menu);
+        menu.findItem(R.id.action_save).setEnabled(!mCommunicating);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.action_save) {
+            onActionSave();
+            return true;
+        } else {
+            boolean handled = super.onOptionsItemSelected(item);
+            if (!handled) {
+                onBackPressed();
+            }
+            return handled;
+        }
+    }
+
+
+    @Override
+    @SuppressWarnings("deprecation")
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        MenuItem saveItem = menu.findItem(R.id.action_save);
+        Drawable tintedIcon = ViewUtils.getTintedIcon(
+                this,
+                R.drawable.ic_checkmark,
+                android.R.color.primary_text_dark);
+        saveItem.setIcon(tintedIcon);
+        return super.onPrepareOptionsMenu(menu);
+    }
+
+
+    protected void setCommunicatingProgress(boolean communicating) {
+        mCommunicating = communicating;
+        if (communicating) {
+            mProgressBar.setVisibility(View.VISIBLE);
+        } else {
+            mProgressBar.setVisibility(View.GONE);
+        }
+        supportInvalidateOptionsMenu();
+    }
+
+    protected abstract void onActionSave();
+
+}
